+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+0.5.18
++++++
+* Fix issue with node config not consuming logging settings
+
+0.5.17
++++++
+* Add parameter '--enable-ultra-ssd' to enable UltraSSD on agent node pool
 
 0.5.17
 +++++